# Replace single TraitDispatchTable with class specific itables to reduce size of binaries