--- a/nativelib/src/main/resources/scala-native/gc/immix_commix/headers/ObjectHeader.h
+++ b/nativelib/src/main/resources/scala-native/gc/immix_commix/headers/ObjectHeader.h
@@ -35,7 +35,7 @@ typedef struct StringObject StringObject;
 
 typedef struct ITableEntry {
     int32_t interfaceId;
-    void *vtable;
+    void **vtable;
 } ITableEntry;
 
 typedef struct Rtti {