--- a/nir/src/main/scala/scala/scalanative/nir/Rt.scala
+++ b/nir/src/main/scala/scala/scalanative/nir/Rt.scala
@@ -36,17 +36,30 @@ private[scalanative] object Rt {
   val ToRawPtrSig = Sig.Method("toRawPtr", Seq(BoxedPtr, Ptr)).mangled
 
   val ClassName = Class.name
-  val ClassIdName = ClassName.member(Sig.Field("id"))
-  val ClassTraitIdName = ClassName.member(Sig.Field("traitId"))
-  val ClassNameName = ClassName.member(Sig.Field("name"))
-  val ClassSizeName = ClassName.member(Sig.Field("size"))
-  val ClassIdRangeUntilName = ClassName.member(Sig.Field("idRangeUntil"))
+  def jlClassFields = Seq(
+    ClassName.member(Sig.Field("id")),
+    ClassName.member(Sig.Field("interfacesCount")),
+    ClassName.member(Sig.Field("interfaces")),
+    ClassName.member(Sig.Field("name")),
+    ClassName.member(Sig.Field("size")),
+    ClassName.member(Sig.Field("idRangeUntil")),
+    ClassName.member(Sig.Field("refFieldOffsets")),
+    ClassName.member(Sig.Field("itablesCount")),
+    ClassName.member(Sig.Field("itables")),
+    ClassName.member(Sig.Field("superClass"))
+  )
 
   val StringName = String.name
   val StringValueName = StringName.member(Sig.Field("value"))
   val StringOffsetName = StringName.member(Sig.Field("offset"))
   val StringCountName = StringName.member(Sig.Field("count"))
   val StringCachedHashCodeName = StringName.member(Sig.Field("cachedHashCode"))
+  def jlStringFields = Seq(
+    StringValueName,
+    StringOffsetName,
+    StringCountName,
+    StringCachedHashCodeName
+  )
 
   val PrimitiveTypes: Seq[Global.Top] = Seq(
     "Byte",