… tools/src/main/scala/scala/scalanative/codegen/Lower.scala
… tools/src/main/scala/scala/scalanative/codegen/TraitDispatchTable.scala
… tools/src/main/scala/scala/scalanative/codegen/TraitsUniverse.scala
… tools/src/main/scala/scala/scalanative/codegen/llvm/MetadataCodeGen.scala
… tools/src/main/scala/scala/scalanative/linker/Infos.scala
… unit-tests/shared/src/test/scala/org/scalanative/testsuite/javalib/lang/ClassTest.scala
… unit-tests/shared/src/test/scala/org/scalanative/testsuite/runtime/LargeTraitsCompositionTest.scala